+++ app/Models/Event.php
@@ -2,10 +2,11 @@
 
 namespace App\Models;
 
-use Cviebrock\EloquentSluggable\Sluggable;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
+use App\Models\EventLink;
 use Illuminate\Database\Eloquent\Model;
+use Cviebrock\EloquentSluggable\Sluggable;
 use Spatie\Activitylog\Traits\LogsActivity;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 /**
  * App\Models\Event
@@ -132,6 +133,11 @@ class Event extends Model
     public function airportArr()
     {
         return $this->hasOne(Airport::class, 'id', 'arr');
+    }
+
+    public function links()
+    {
+        return $this->hasMany(EventLink::class);
     }
 
     public function faqs()